requirements.txt
requests
Pillow
boto3
Flask
psycopg2